% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_uv.R
\name{correct.uv}
\alias{correct.uv}
\title{U and V correction}
\usage{
correct.uv(U, V, sigma)
}
\arguments{
\item{U}{Estimated latent factor matrix.}

\item{V}{Estimated factor loadings.}

\item{sigma}{Estimated variance of latent factor.}
}
\value{
Corrected U and V
}
\description{
Rotates U and V such that the covariance of U is diagonal, and V is upper-triangular with a positive diagonal
}
