% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimation}
\alias{estimation}
\title{Estimate process for specific candidate p and seed}
\usage{
estimation(Y, X, logs, p, seed, maxIter, alpha, tol1, tol2, out.dir)
}
\arguments{
\item{Y}{Concatenated count matrix.}

\item{X}{Covariate information.}

\item{logs}{Log library size factor using whole gene list instead of only highly variable genes.}

\item{p}{A specific hyper parameter p.}

\item{seed}{A specific initialization seed.}

\item{maxIter}{The number of maximum iteration. (default 5000)}

\item{alpha}{Step size. (default 0.01)}

\item{tol1}{The stopping rule for RMSE. (default 1e-5)}

\item{tol2}{The stopping rule for mean RMSE. (default 1.5e-5)}

\item{out.dir}{Output directory.}
}
\value{
A list of estimates.

#' @references
Kidzinski, L., Hui, F.K., Warton, D.I. and Hastie, T.J., 2022. Generalized Matrix Factorization: efficient algorithms for fitting generalized linear latent variable models to large data arrays.
Journal of Machine Learning Research, 23(291), pp.1-29.
}
\description{
This function input the concatenated count matrix, covariate information and log library size factor,
and outputs a list of
}
