% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{final_result}
\alias{final_result}
\title{Final results from scINSIGHT2}
\format{
An object of class \code{list} of length 9.
}
\usage{
final_result
}
\description{
Final results from scINSIGHT2
}
\keyword{datasets}
