% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_index.R
\name{integration_score}
\alias{integration_score}
\title{Integration index}
\usage{
integration_score(matrix, s = 50, l = 500, covariates)
}
\arguments{
\item{matrix}{The normalized latent variables.}

\item{s}{The number of cells that are closest neighbors of cell $i$.}

\item{l}{The number of cells that are in a larger neighborhood of cell $i$.}

\item{covariates}{A list of covariate vectors.}
}
\value{
A list of integration index scores.
}
\description{
Calculate integration index for each covariate
}
