% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{libsize}
\alias{libsize}
\title{A vector of library size for each cell.}
\format{
An object of class \code{numeric} of length 2000.
}
\usage{
libsize
}
\description{
A vector of library size for each cell.
}
\keyword{datasets}
