% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scINSIGHT2.R
\name{scINSIGHT2_estimate}
\alias{scINSIGHT2_estimate}
\title{Main unction in scINSIGHT2 package}
\usage{
scINSIGHT2_estimate(
  object,
  libsize = NULL,
  p_candidate = NULL,
  B = 1,
  maxIter = 5000,
  alpha = 0.01,
  tol1 = 1e-05,
  tol2 = 1.5e-05,
  num.cores = 1,
  out.dir = NULL
)
}
\arguments{
\item{object}{A Seurat object containing count matrix Y and covariate information X in meta.data.}

\item{libsize}{library size of all the cells if provided.}

\item{p_candidate}{The vector of candidate p values.}

\item{B}{Number of different initialization seeds to be selected from. The candidates are 1, ..., B. (default 1)}

\item{maxIter}{The number of maximum iteration. (default 5000)}

\item{alpha}{Step size.(default 0.01)}

\item{tol1}{The stopping rule for RMSE. (default 1e-5)}

\item{tol2}{The stopping rule for mean RMSE. (default 1.5e-5)}

\item{num.cores}{Number of cores used for estimation. (default 1)}

\item{out.dir}{Output directory.}
}
\value{
A list of estimates.
}
\description{
To estimate latent factors, corresponding loadings and coefficients of covariates.
}
