% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{selection}
\alias{selection}
\title{Selection of latent dimension p and seed d}
\usage{
selection(U_all, individual, n_cell, p_candidate, seeds)
}
\arguments{
\item{U_all}{The estimation of latent factors U with each candidate p and seed pair.}

\item{individual}{A vector indicating different samples.}

\item{n_cell}{Total number of cells in the count matrix.}

\item{p_candidate}{The candidate p values.}

\item{seeds}{Candidate initialization seeds.}
}
\value{
A list of selected latent dimension p and seed.
}
\description{
Select latent dimension p and the corresponding initialization seed d to finalize the estimation
}
