% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_object}
\alias{sim_object}
\title{A Seurat object containing count matrix and meta.data.}
\format{
An object of class \code{Seurat} with 2000 rows and 5000 columns.
}
\usage{
sim_object
}
\description{
A Seurat object containing count matrix and meta.data.
}
\keyword{datasets}
